+++ examples/middlewares/middlewares/log-access.ts
@@ -1,5 +1,5 @@
 import { Service } from "typedi";
-import { MiddlewareInterface, NextFn, ActionData } from "../../../src";
+import { MiddlewareInterface, NextFn, ResolverData } from "../../../src";
 
 import { Context } from "../context";
 import { Logger } from "../logger";
@@ -8,7 +8,7 @@ import { Logger } from "../logger";
 export class LogAccessMiddleware implements MiddlewareInterface<Context> {
   constructor(private readonly logger: Logger) {}
 
-  async use({ context, info }: ActionData<Context>, next: NextFn) {
+  async use({ context, info }: ResolverData<Context>, next: NextFn) {
     const username: string = context.username || "guest";
     this.logger.log(`Logging access: ${username} -> ${info.parentType.name}.${info.fieldName}`);
     return next();