+++ src/types/filter-action-data.ts
@@ -2,7 +2,7 @@ import { GraphQLResolveInfo } from "graphql";
 
 import { ArgsDictionary } from "./action-data";
 
-export interface FilterActionData<TPayload = any, TArgs = ArgsDictionary, TContext = {}> {
+export interface ResolverFilterData<TPayload = any, TArgs = ArgsDictionary, TContext = {}> {
   payload: TPayload;
   args: TArgs;
   context: TContext;