+++ src/types/index.ts
@@ -1,2 +1,6 @@
-export { ActionData } from "./action-data";
-export { FilterActionData } from "./filter-action-data";
+export { ResolverData } from "./action-data";
+export { ResolverFilterData } from "./filter-action-data";
+
+// TODO: remove depreciated aliases
+export { ResolverData as ActionData } from "./action-data";
+export { ResolverFilterData as FilterActionData } from "./filter-action-data";